+++ src/main/java/com/netgrif/application/engine/elastic/web/ElasticController.java
@@ -2,6 +2,8 @@ package com.netgrif.application.engine.elastic.web;
 
 import com.netgrif.application.engine.auth.domain.LoggedUser;
 import com.netgrif.application.engine.elastic.service.ReindexingTask;
+import com.netgrif.application.engine.elastic.service.interfaces.IElasticIndexService;
+import com.netgrif.application.engine.elastic.web.requestbodies.IndexParams;
 import com.netgrif.application.engine.workflow.service.CaseSearchService;
 import com.netgrif.application.engine.workflow.service.interfaces.IWorkflowService;
 import com.netgrif.application.engine.workflow.web.responsebodies.MessageResource;
@@ -20,10 +22,7 @@ import org.springframework.hateoas.MediaTypes;
 import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.Locale;
 import java.util.Map;
@@ -49,6 +48,9 @@ public class ElasticController {
     @Autowired
     private ReindexingTask reindexingTask;
 
+    @Autowired
+    private IElasticIndexService indexService;
+
     @Value("${spring.data.elasticsearch.reindexExecutor.size:20}")
     private int pageSize;
 
@@ -69,16 +71,35 @@ public class ElasticController {
             if (count == 0) {
                 log.info("No cases to reindex");
             } else {
-                long numOfPages = (long) ((count / pageSize) + 1);
-                log.info("Reindexing cases: " + numOfPages + " pages");
+                long numOfPages = (count / pageSize) + 1;
+                log.info("Reindexing cases: {} pages", numOfPages);
 
                 for (int page = 0; page < numOfPages; page++) {
-                    log.info("Indexing page " + (page + 1));
+                    log.info("Indexing page {}", (page + 1));
                     Predicate predicate = searchService.buildQuery(searchBody, user, locale);
                     reindexingTask.forceReindexPage(predicate, page, numOfPages);
                 }
             }
 
+            return MessageResource.successMessage("Success");
+        } catch (Exception e) {
+            log.error("Could not index: ", e);
+            return MessageResource.errorMessage(e.getMessage());
+        }
+    }
+
+    @PreAuthorize("@authorizationService.hasAuthority('ADMIN')")
+    @Operation(summary = "Reindex all or stale cases with bulk index",
+            description = "Reindex all or stale cases (specified by IndexParams.indexAll param) with bulk index. Caller must have the ADMIN role",
+            security = {@SecurityRequirement(name = "BasicAuth")})
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "OK"),
+            @ApiResponse(responseCode = "403", description = "Caller doesn't fulfill the authorisation requirements"),
+    })
+    @PostMapping(value = "/reindex/bulk", produces = MediaType.APPLICATION_JSON_VALUE)
+    public MessageResource bulkReindex(IndexParams indexParams) {
+        try {
+            indexService.bulkIndex(indexParams.isIndexAll(), null, indexParams.getCaseBatchSize(), indexParams.getTaskBatchSize());
             return MessageResource.successMessage("Success");
         } catch (Exception e) {
             log.error("Could not index: ", e);