+++ src/main/java/com/netgrif/application/engine/workflow/domain/DataField.java
@@ -14,6 +14,7 @@ import com.querydsl.core.annotations.QueryType;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.annotation.Transient;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
@@ -65,6 +66,11 @@ public class DataField implements Referencable, Serializable {
     @Setter
     private Component component;
 
+    @Transient
+    @Getter
+    @Setter
+    private boolean changed = false;
+
     public DataField() {
         behavior = new HashMap<>();
         dataRefComponents = new HashMap<>();
@@ -81,13 +87,27 @@ public class DataField implements Referencable, Serializable {
     }
 
     public void setValue(Object value) {
+        setValue(value, true);
+    }
+
+    public void setValue(Object value, boolean trackChange) {
         this.value = value;
         update();
+        if (trackChange) {
+            changed();
+        }
     }
 
     public void setChoices(Set<I18nString> choices) {
+        setChoices(choices, true);
+    }
+
+    public void setChoices(Set<I18nString> choices, boolean trackChange) {
         this.choices = choices;
         update();
+        if (trackChange) {
+            changed();
+        }
     }
 
     public void setAllowedNets(List<String> allowedNets) {
@@ -101,8 +121,15 @@ public class DataField implements Referencable, Serializable {
     }
 
     public void setOptions(Map<String, I18nString> options) {
+        setOptions(options, true);
+    }
+
+    public void setOptions(Map<String, I18nString> options, boolean trackChange) {
         this.options = options;
         update();
+        if (trackChange) {
+            changed();
+        }
     }
 
     public void setValidations(List<Validation> validations) {
@@ -208,6 +235,10 @@ public class DataField implements Referencable, Serializable {
 
     private void update() {
         version++;
+    }
+
+    private void changed() {
+        changed = true;
     }
 
     public boolean isNewerThen(DataField other) {