+++ src/test/groovy/com/netgrif/application/engine/workflow/TaskControllerTest.groovy
@@ -1,5 +1,7 @@
 package com.netgrif.application.engine.workflow
 
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.node.ObjectNode
 import com.netgrif.application.engine.TestHelper
 import com.netgrif.application.engine.auth.domain.Authority
 import com.netgrif.application.engine.auth.domain.User
@@ -9,7 +11,6 @@ import com.netgrif.application.engine.auth.service.interfaces.IUserService
 import com.netgrif.application.engine.elastic.service.interfaces.IElasticTaskService
 import com.netgrif.application.engine.petrinet.domain.PetriNet
 import com.netgrif.application.engine.petrinet.domain.VersionType
-import com.netgrif.application.engine.petrinet.domain.dataset.FileFieldValue
 import com.netgrif.application.engine.petrinet.domain.dataset.FileListFieldValue
 import com.netgrif.application.engine.petrinet.domain.roles.ProcessRole
 import com.netgrif.application.engine.petrinet.service.ProcessRoleService
@@ -18,12 +19,12 @@ import com.netgrif.application.engine.startup.ImportHelper
 import com.netgrif.application.engine.startup.SuperCreator
 import com.netgrif.application.engine.utils.FullPageRequest
 import com.netgrif.application.engine.workflow.domain.Case
+import com.netgrif.application.engine.workflow.domain.DataField
 import com.netgrif.application.engine.workflow.domain.Task
 import com.netgrif.application.engine.workflow.service.TaskSearchService
 import com.netgrif.application.engine.workflow.service.TaskService
 import com.netgrif.application.engine.workflow.service.interfaces.IDataService
 import com.netgrif.application.engine.workflow.service.interfaces.IWorkflowService
-import com.netgrif.application.engine.workflow.web.PublicTaskController
 import com.netgrif.application.engine.workflow.web.TaskController
 import com.netgrif.application.engine.workflow.web.WorkflowController
 import com.netgrif.application.engine.workflow.web.requestbodies.TaskSearchRequest
@@ -87,7 +88,9 @@ class TaskControllerTest {
     @Autowired
     private TaskController taskController
 
-    private PetriNet net
+    private PetriNet allDataNet
+
+    private PetriNet setDataNet
 
     private Case useCase
 
@@ -106,7 +109,7 @@ class TaskControllerTest {
                 state: UserState.ACTIVE,
                 authorities: [authorityService.getOrCreate(Authority.user)] as Set<Authority>,
                 processRoles: [] as Set<ProcessRole>))
-        importNet()
+        importNets()
     }
 
     @Test
@@ -118,7 +121,7 @@ class TaskControllerTest {
 
     @Test
     void testDeleteFile() {
-        Case testCase = helper.createCase("My case", net)
+        Case testCase = helper.createCase("My case", allDataNet)
         String taskId = testCase.tasks.find {it.transition == "1"}.task
 
         dataService.saveFile(taskId, "file", new MockMultipartFile("test", new byte[] {}))
@@ -132,7 +135,7 @@ class TaskControllerTest {
 
     @Test
     void testDeleteFileByName() {
-        Case testCase = helper.createCase("My case", net)
+        Case testCase = helper.createCase("My case", allDataNet)
         String taskId = testCase.tasks.find {it.transition == "1"}.task
 
         dataService.saveFiles(taskId, "fileList", new MockMultipartFile[] {new MockMultipartFile("test", "test", null, new byte[] {})})
@@ -142,6 +145,79 @@ class TaskControllerTest {
         taskController.deleteNamedFile(taskId, "fileList", "test")
         testCase = workflowService.findOne(testCase.stringId)
         assert ((FileListFieldValue) testCase.dataSet["fileList"].value).namesPaths == null || ((FileListFieldValue) testCase.dataSet["fileList"].value).namesPaths.size() == 0
+    }
+
+    @Test
+    void testSetDataFieldTypeRestriction() {
+        Case testCase = helper.createCase("My case", setDataNet)
+        String taskId = testCase.tasks.find { it.transition == "testSetDataFieldTypeRestriction" }.task
+
+        ObjectNode dataSet = populateNestedDataset([(taskId):["taskRef_0": ["type": "taskRef", "value": [taskId]]]])
+        def response = taskController.setData(taskId, dataSet, Locale.default)
+        assert response != null && response.content.outcome != null
+        assert response.content.outcome.changedFields.changedFields.isEmpty()
+        assert ((List<String>) workflowService.findOne(testCase.stringId).getDataField("taskRef_0").getValue()).isEmpty()
+
+        dataSet = populateNestedDataset([(taskId):["caseRef_0": ["type": "caseRef", "value": [testCase.stringId]]]])
+        response = taskController.setData(taskId, dataSet, Locale.default)
+        assert response != null && response.content.outcome != null
+        assert response.content.outcome.changedFields.changedFields.isEmpty()
+        assert ((List<String>) workflowService.findOne(testCase.stringId).getDataField("caseRef_0").getValue()).isEmpty()
+    }
+
+    @Test
+    void testSetDataVisibleField() {
+        Case testCase = helper.createCase("My case", setDataNet)
+        String taskId = testCase.tasks.find { it.transition == "data" }.task
+
+        ObjectNode dataSet = populateNestedDataset([(taskId):["text_1": ["type": "text", "value": "awd"]]])
+        def response = taskController.setData(taskId, dataSet, Locale.default)
+        assert response != null && response.content.outcome == null
+        assert response.content.error != null
+
+        // todo: test visible behavior based on parent taskRef behavior
+    }
+
+    @Test
+    void testSetDataNestedTaskRefRestrictions() {
+        Case testCase1 = helper.createCase("testCase1", setDataNet)
+        String taskId = testCase1.tasks.find { it.transition == "testSetDataNestedTaskRefRestrictions" }.task
+        Case testCase2 = helper.createCase("testCase2", setDataNet)
+        Case testCase3 = helper.createCase("testCase3", setDataNet)
+
+        DataField case1DataField = testCase1.getDataField("taskRef_0")
+        case1DataField.setValue(List.of(testCase2.tasks.find { it.transition == "testSetDataNestedTaskRefRestrictions" }.task))
+        workflowService.save(testCase1)
+
+        DataField case2DataField = testCase2.getDataField("taskRef_0")
+        case2DataField.setValue(List.of(testCase3.tasks.find { it.transition == "data" }.task))
+        workflowService.save(testCase2)
+
+        String nestedOtherTaskId = testCase2.tasks.find { it.transition == "data" }.task
+        ObjectNode dataSet = populateNestedDataset([(nestedOtherTaskId):["text_0": ["type": "text", "value": "awd"]]])
+        def response = taskController.setData(taskId, dataSet, Locale.default)
+        assert response != null && response.content.outcome != null
+        assert response.content.outcome.changedFields.changedFields.isEmpty()
+        assert workflowService.findOne(testCase2.stringId).getDataField("text_0").getValue() == null
+
+        String nestedTaskId = testCase3.tasks.find { it.transition == "data" }.task
+        dataSet = populateNestedDataset([(nestedTaskId):["text_0": ["type": "text", "value": "awd"]]])
+        response = taskController.setData(taskId, dataSet, Locale.default)
+        assert response != null && response.content.outcome != null
+        assert !response.content.outcome.changedFields.changedFields.isEmpty()
+        assert workflowService.findOne(testCase3.stringId).getDataField("text_0").getValue() == "awd"
+    }
+
+    @Test
+    void testSetDataNonReferencedField() {
+        Case testCase = helper.createCase("My case", setDataNet)
+        String taskId = testCase.tasks.find { it.transition == "testSetDataNonReferencedField" }.task
+
+        ObjectNode dataSet = populateNestedDataset([(taskId):["text_1": ["type": "text", "value": "awd"]]])
+        def response = taskController.setData(taskId, dataSet, Locale.default)
+
+        assert response != null && response.content.outcome == null
+        assert response.content.error != null
     }
 
 
@@ -167,15 +243,19 @@ class TaskControllerTest {
         assert !findTasksByMongo().empty
     }
 
-    void importNet() {
-        PetriNet netOptional = helper.createNet("all_data_refs.xml", VersionType.MAJOR).get()
-        assert netOptional != null
-        net = netOptional
+    void importNets() {
+        PetriNet allDataNet = helper.createNet("all_data_refs.xml", VersionType.MAJOR).get()
+        assert allDataNet != null
+        this.allDataNet = allDataNet
+
+        PetriNet setDataNet = helper.createNet("task_controller_set_data.xml", VersionType.MAJOR).get()
+        assert setDataNet != null
+        this.setDataNet = setDataNet
     }
 
     void createCase() {
         useCase = null
-        useCase = helper.createCase("My case", net)
+        useCase = helper.createCase("My case", allDataNet)
         assert useCase != null
     }
 
@@ -203,7 +283,7 @@ class TaskControllerTest {
     }
 
     void setUserRole() {
-        List<ProcessRole> roles = processRoleService.findAll(net.stringId)
+        List<ProcessRole> roles = processRoleService.findAll(allDataNet.stringId)
 
         for (ProcessRole role : roles) {
             if (role.importId == "process_role") {
@@ -218,5 +298,11 @@ class TaskControllerTest {
         taskSearchRequestList.add(new TaskSearchRequest())
         Page<Task> tasks = taskService.search(taskSearchRequestList, new FullPageRequest(), userService.findByEmail(DUMMY_USER_MAIL, false).transformToLoggedUser(), new Locale("en"), false)
         return tasks
+    }
+
+    static ObjectNode populateNestedDataset(Map<String, Map<String, Map<String, Object>>> data) {
+        ObjectMapper mapper = new ObjectMapper()
+        String json = mapper.writeValueAsString(data)
+        return mapper.readTree(json) as ObjectNode
     }
 }